// src/domain/types/project-status.type.ts
export enum ProjectStatus {
  Cancelled = 'Cancelled',
  Rejected = 'Rejected',
  RejectedAfterDueDiligence = 'RejectedAfterDueDiligence',
  PendingFirstContract = 'PendingFirstContract',
  PendingToApprove = 'PendingToApprove',
  PendingClosed = 'PendingClosed',
  Expired = 'Expired',
  Approved = 'Approved',
  AgreementClosed = 'AgreementClosed',
  TechnicalScoringAssets10 = 'TechnicalScoringAssets10',
  UnderContractWithOwner = 'UnderContractWithOwner',
  FirstFundingIncome = 'FirstFundingIncome',
  DueDiligence = 'DueDiligence',
  ContractWithCapital = 'ContractWithCapital',
  FundingAgreement = 'FundingAgreement',
  Escrow = 'Escrow',
  PermittingAndStructuring = 'PermittingAndStructuring',
  PermitsApproved = 'PermitsApproved',
  FinalClosing = 'FinalClosing',
  NotEconomicallyViable = 'NotEconomicallyViable',

  // new status
  PassedCriteria = 'PassedCriteria',
  PendingResources = 'PendingResources',
  TechnicalReview = 'TechnicalReview',
  FinancingReview = 'FinancingReview',
  PromiseSignedWithPropietary = 'PromiseSignedWithPropietary',
  TitleNDAStudies = 'TitleNDAStudies',
  PromiseAgreementWithDeveloper = 'PromiseAgreementWithDeveloper',
  AnteProjectApproved = 'AnteProjectApproved',
  ApprovedTitles = 'ApprovedTitles',
  FinalPromiseAndPayment = 'FinalPromiseAndPayment',
  Closed = 'Closed',
}